+++ fcli-core/fcli-ssc/src/main/resources/com/fortify/cli/ssc/i18n/SSCMessages.properties
@@ -232,8 +232,8 @@ fcli.ssc.appversion.create.copy-options = Comma separated list of elements to co
   Allowed values: ${COMPLETION-CANDIDATES}.
 fcli.ssc.appversion.delete.usage.header = Delete an application version.
 fcli.ssc.appversion.copy-from.usage.header = Copy from options
-fcli.ssc.appversion.copy-state.usage.header = Copy application version state.
-fcli.ssc.appversion.copy-state.usage.description = Copy application version state from another application version
+fcli.ssc.appversion.copy-state.usage.header = (PREVIEW) Copy application version state.
+fcli.ssc.appversion.copy-state.usage.description = Copy application version state from another application version. \nNote that this command may potentially have unintended side effects when copying state to application versions that already contain artifacts.
 fcli.ssc.appversion.download-state.usage.header = Download application version state.
 fcli.ssc.appversion.download-state.usage.description = Download application version state artifact. See 'fcli ssc artifact download' for downloading individual artifacts.
 fcli.ssc.appversion.download-state.file = Optional output file path.
@@ -271,7 +271,8 @@ fcli.ssc.appversion.update.description = Update application version description.
 fcli.ssc.appversion.resolver.name = Application and version name.
 fcli.ssc.appversion.resolver.nameOrId = Application version id or <application>:<version> name.
 fcli.ssc.appversion.resolver.copy-from.nameOrId = Copy FROM application version: \nid or <application>:<version> name.
-  
+fcli.ssc.appversion.resolver.copy-to.nameOrId = Copy TO application version: \nid or <application>:<version> name.
+
 # fcli ssc artifact
 fcli.ssc.artifact.usage.header = Manage SSC artifacts.
 fcli.ssc.artifact.approve.usage.header = Approve an artifact.
